Refactor lineisland into its own file; 15x15 line solving cases
