+++ nonosolve2/tests/lines_15x15.py
@@ -1,0 +1,27 @@
+lines = [
+    {
+        'line': [' ', ' ', ' ', ' ', ' ', ' ', ' ', 'X', 'X', ' ', ' ', ' ', ' ', ' ', ' '],
+        'islands': [5],
+        'result': """
+|  |  |  |  |  |  |  |X |X |  |  |  |  |  |  | 5
+ *  *  *  *  *  *  *        *  *  *  *  *  *   5
+"""
+    },
+    {
+        'line': [' ', ' ', 'X', ' ', ' ', ' ', ' ', 'X', 'X', ' ', ' ', ' ', ' ', ' ', ' '],
+        'islands': [5],
+        'result': """
+|X |X |X |X |X |X |X |X |X |  |O |O |O |O |  | 5
+                            *  *  *  *  *  *   5
+"""
+    },
+    {
+        'line': [' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' '],
+        'islands': [13, 1],
+        'result': """
+|O |O |O |O |O |O |O |O |O |O |O |O |O |X |O | [13] [1]
+ *  *  *  *  *  *  *  *  *  *  *  *  *         13
+                                           *   1
+"""
+    }
+]
