+++ nonosolve2/tests/test_singlesolveline15x15.py
@@ -1,0 +1,7 @@
+import unittest
+
+class TestSingleSolveLine15x15(unittest.TestCase):
+    pass
+
+if __name__ == '__main__':
+    unittest.main()
